refactor(atomic-a11y): extract shared types, guards, constants and WCAG data modules
- Create src/shared/types.ts with unified CriterionLevel superset
- Create src/shared/guards.ts with deduplicated isRecord/isA11yReport
- Create src/shared/constants.ts with all shared constants
- Create src/shared/file-utils.ts with file I/O utilities
- Create src/data/wcag-criteria.ts with 56 WCAG criteria definitions
- Create src/data/axe-rule-mappings.ts with axe-to-WCAG mappings
- Create src/data/criterion-metadata.ts with criterion metadata
- Update vitest-a11y-reporter.ts to use shared modules (~200 lines removed)
- Update json-to-openacr.ts to use shared modules (~371 lines removed)
- Update merge-shards.ts to use shared modules (16 lines removed)

Total extraction: 776 lines into 7 new focused modules
Zero code duplication for guards/types/constants achieved
