+++ packages/atomic-a11y/src/shared/constants.ts
@@ -1,0 +1,89 @@
+export const DEFAULT_A11Y_REPORT_OUTPUT_DIR = 'a11y/reports';
+export const DEFAULT_A11Y_REPORT_FILENAME = 'a11y-report.json';
+export const DEFAULT_WCAG_22_AA_CRITERIA_COUNT = 50;
+export const BASELINE_FILE_PATTERN =
+  /^manual-audit-(?!.*-violations)([\w-]+)\.json$/;
+export const DELTA_PATTERN = /^delta-(\d{4}-\d{2}-\d{2})-([\w-]+)\.json$/;
+export const VALID_STATUSES = new Set([
+  'pass',
+  'fail',
+  'partial',
+  'not-applicable',
+]);
+export const VALID_WCAG_KEYS = new Set([
+  '1.1.1',
+  '1.2.1',
+  '1.2.2',
+  '1.2.3',
+  '1.2.4',
+  '1.2.5',
+  '1.3.1',
+  '1.3.2',
+  '1.3.3',
+  '1.3.4',
+  '1.3.5',
+  '1.4.1',
+  '1.4.2',
+  '1.4.3',
+  '1.4.4',
+  '1.4.5',
+  '1.4.10',
+  '1.4.11',
+  '1.4.12',
+  '1.4.13',
+  '2.1.1',
+  '2.1.2',
+  '2.1.4',
+  '2.2.1',
+  '2.2.2',
+  '2.3.1',
+  '2.4.1',
+  '2.4.2',
+  '2.4.3',
+  '2.4.4',
+  '2.4.5',
+  '2.4.6',
+  '2.4.7',
+  '2.4.11',
+  '2.5.1',
+  '2.5.2',
+  '2.5.3',
+  '2.5.4',
+  '2.5.7',
+  '2.5.8',
+  '3.1.1',
+  '3.1.2',
+  '3.2.1',
+  '3.2.2',
+  '3.2.3',
+  '3.2.4',
+  '3.2.6',
+  '3.3.1',
+  '3.3.2',
+  '3.3.3',
+  '3.3.4',
+  '3.3.7',
+  '3.3.8',
+  '4.1.1',
+  '4.1.2',
+  '4.1.3',
+]);
+export const ALL_AI_CRITERIA = [
+  '1.1.1',
+  '1.2.2',
+  '1.3.1',
+  '1.3.5',
+  '1.4.1',
+  '1.4.3',
+  '1.4.4',
+  '2.1.1',
+  '2.2.1',
+  '2.4.1',
+  '2.4.2',
+  '2.4.3',
+  '2.4.4',
+  '2.5.8',
+  '3.1.1',
+  '4.1.1',
+  '4.1.2',
+];
